fixes and more fleshed out code
- Fixed Extended Bitvectors as well as use
- Fixed string compare
- Added some Macros such as URANGE, UMIN and UMAX
- Flesh out damage, devote, supplicate, empty, fill, addaffect

+++ Source/SmaugCS.Common/StringExtensions.cs
@@ -110,7 +110,7 @@ namespace SmaugCS.Common
 
             foreach (string number in numbers)
             {
-                if (!ulong.TryParse(number, out ulong num))
+                if (!int.TryParse(number, out int num))
                     continue;
 
                 bit.SetBit(num);
